+++ crates/cxx-qt-gen/src/syntax/attribute.rs
@@ -97,14 +97,14 @@ mod tests {
         let module: ItemMod = parse_quote! {
             #[qinvokable]
             #[cxx_qt::bridge]
-            #[cxx_qt::qsignals(MyObject)]
+            #[cxx_qt::object(MyObject)]
             #[cxx_qt::bridge(namespace = "my::namespace")]
             mod module;
         };
 
         assert!(attribute_find_path(&module.attrs, &["qinvokable"]).is_some());
         assert!(attribute_find_path(&module.attrs, &["cxx_qt", "bridge"]).is_some());
-        assert!(attribute_find_path(&module.attrs, &["cxx_qt", "qsignals"]).is_some());
+        assert!(attribute_find_path(&module.attrs, &["cxx_qt", "object"]).is_some());
         assert!(attribute_find_path(&module.attrs, &["cxx_qt", "missing"]).is_none());
     }
 
@@ -113,7 +113,7 @@ mod tests {
         let module: ItemMod = parse_quote! {
             #[qinvokable]
             #[cxx_qt::bridge]
-            #[cxx_qt::qsignals(MyObject)]
+            #[cxx_qt::object(MyObject)]
             #[cxx_qt::bridge(namespace = "my::namespace")]
             #[cxx_qt::list(A, B, C)]
             #[cxx_qt::bridge(a = "b", namespace = "my::namespace")]